refactor: community 관련 Type 클래스 패키지 이동
- BoardCode를 community.board.domain 으로
- PostCategory를 community.post.domain 으로
- RedisConstants 를 community.post.service 로
