+++ src/main/java/com/example/solidconnection/community/board/domain/BoardCode.java
@@ -1,4 +1,4 @@
-package com.example.solidconnection.type;
+package com.example.solidconnection.community.board.domain;
 
 public enum BoardCode {
     EUROPE, AMERICAS, ASIA, FREE;
