+++ src/main/java/com/example/solidconnection/community/post/domain/PostCategory.java
@@ -1,0 +1,5 @@
+package com.example.solidconnection.community.post.domain;
+
+public enum PostCategory {
+    전체, 자유, 질문
+}
