+++ src/main/java/com/example/solidconnection/community/post/dto/PostCreateRequest.java
@@ -3,7 +3,7 @@ package com.example.solidconnection.community.post.dto;
 import com.example.solidconnection.community.board.domain.Board;
 import com.example.solidconnection.community.post.domain.Post;
 import com.example.solidconnection.siteuser.domain.SiteUser;
-import com.example.solidconnection.type.PostCategory;
+import com.example.solidconnection.community.post.domain.PostCategory;
 import jakarta.validation.constraints.NotBlank;
 import jakarta.validation.constraints.NotNull;
 import jakarta.validation.constraints.Size;
