+++ src/main/java/com/example/solidconnection/community/post/service/PostQueryService.java
@@ -15,8 +15,8 @@ import com.example.solidconnection.community.post.repository.PostRepository;
 import com.example.solidconnection.service.RedisService;
 import com.example.solidconnection.siteuser.domain.SiteUser;
 import com.example.solidconnection.siteuser.dto.PostFindSiteUserResponse;
-import com.example.solidconnection.type.BoardCode;
-import com.example.solidconnection.type.PostCategory;
+import com.example.solidconnection.community.board.domain.BoardCode;
+import com.example.solidconnection.community.post.domain.PostCategory;
 import com.example.solidconnection.util.RedisUtils;
 import lombok.RequiredArgsConstructor;
 import org.apache.commons.lang3.EnumUtils;
