--- src/main/java/com/example/solidconnection/type/PostCategory.java
@@ -1,5 +1,0 @@
-package com.example.solidconnection.type;
-
-public enum PostCategory {
-    전체, 자유, 질문
-}
