+++ src/test/java/com/example/solidconnection/community/post/service/PostCommandServiceTest.java
@@ -18,7 +18,7 @@ import com.example.solidconnection.siteuser.domain.SiteUser;
 import com.example.solidconnection.siteuser.fixture.SiteUserFixture;
 import com.example.solidconnection.support.integration.BaseIntegrationTest;
 import com.example.solidconnection.type.ImgType;
-import com.example.solidconnection.type.PostCategory;
+import com.example.solidconnection.community.post.domain.PostCategory;
 import com.example.solidconnection.util.RedisUtils;
 import jakarta.transaction.Transactional;
 import org.junit.jupiter.api.BeforeEach;
