+++ src/test/java/com/example/solidconnection/concurrency/PostViewCountConcurrencyTest.java
@@ -8,8 +8,8 @@ import com.example.solidconnection.service.RedisService;
 import com.example.solidconnection.siteuser.domain.SiteUser;
 import com.example.solidconnection.siteuser.fixture.SiteUserFixture;
 import com.example.solidconnection.support.TestContainerSpringBootTest;
-import com.example.solidconnection.type.PostCategory;
-import com.example.solidconnection.util.RedisUtils;
+import com.example.solidconnection.community.post.domain.PostCategory;
+\import com.example.solidconnection.util.RedisUtils;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
@@ -21,7 +21,7 @@ import java.util.concurrent.ExecutorService;
 import java.util.concurrent.Executors;
 import java.util.concurrent.TimeUnit;
 
-import static com.example.solidconnection.type.RedisConstants.VALIDATE_VIEW_COUNT_TTL;
+import static com.example.solidconnection.community.post.service.RedisConstants.*;
 import static org.junit.jupiter.api.Assertions.assertEquals;
 
 @TestContainerSpringBootTest
